use button tabs on space + project page
