+++ apps/client/package.json
@@ -23,6 +23,7 @@
     "@prisma/client": "^4.9.0",
     "@radix-ui/react-context-menu": "^2.1.1",
     "@radix-ui/react-dialog": "^1.0.2",
+    "@radix-ui/react-tabs": "^1.0.2",
     "@radix-ui/react-toggle-group": "^1.0.2",
     "@radix-ui/react-tooltip": "^1.0.3",
     "@rainbow-me/rainbowkit": "^0.8.1",
