+++ apps/client/src/home/NavbarLayout/NavbarLayout.tsx
