+++ apps/client/src/home/NavbarLayout/ProfileMenu.tsx
@@ -1,10 +1,10 @@
 import Link from "next/link";
 import { MdLogout, MdOutlinePersonOutline, MdOutlineSettings } from "react-icons/md";
 
-import { useLogout } from "../../../client/auth/useLogout";
-import { useSession } from "../../../client/auth/useSession";
-import { trpc } from "../../../client/trpc";
-import { numberToHexDisplay } from "../../../utils/numberToHexDisplay";
+import { useLogout } from "../../client/auth/useLogout";
+import { useSession } from "../../client/auth/useSession";
+import { trpc } from "../../client/trpc";
+import { numberToHexDisplay } from "../../utils/numberToHexDisplay";
 import ProfileMenuButton from "./ProfileMenuButton";
 
 interface Props {
