+++ apps/client/src/pages/settings.tsx
@@ -8,8 +8,8 @@ import { useSession } from "../client/auth/useSession";
 import { trpc } from "../client/trpc";
 import { cropImage } from "../editor/utils/cropImage";
 import { env } from "../env/client.mjs";
-import { getNavbarLayout } from "../home/layouts/NavbarLayout/NavbarLayout";
 import MetaTags from "../home/MetaTags";
+import { getNavbarLayout } from "../home/NavbarLayout/NavbarLayout";
 import Button from "../ui/Button";
 import ImageInput from "../ui/ImageInput";
 import Spinner from "../ui/Spinner";
