+++ apps/client/src/ui/ButtonTabs.tsx
@@ -1,0 +1,40 @@
+import * as Tabs from "@radix-ui/react-tabs";
+import { useState } from "react";
+
+interface Props {
+  titles: string[];
+  children: React.ReactNode;
+}
+
+export default function ButtonTabs({ titles, children }: Props) {
+  const [value, setValue] = useState(titles[0]);
+
+  return (
+    <Tabs.Root
+      defaultValue={titles[0]}
+      value={value}
+      onValueChange={setValue}
+      className="space-y-4"
+    >
+      {titles.length > 1 && (
+        <Tabs.List className="flex space-x-2">
+          {titles.map((title) => (
+            <Tabs.Trigger
+              key={title}
+              value={title}
+              className={`flex-1 rounded-lg py-1 text-lg font-bold transition active:opacity-80 ${
+                value === title ? "bg-neutral-200 hover:bg-neutral-300" : " hover:bg-neutral-200"
+              }`}
+            >
+              {title}
+            </Tabs.Trigger>
+          ))}
+        </Tabs.List>
+      )}
+
+      {children}
+    </Tabs.Root>
+  );
+}
+
+export const TabContent = Tabs.Content;
